+++ include/solvers/fluid_dynamics_matrix_free_operators.h
@@ -141,7 +141,8 @@ boundary_condition_requires_face_assembly(
 {
   if (boundary_type == BoundaryConditions::BoundaryType::function_weak ||
       boundary_type == BoundaryConditions::BoundaryType::outlet ||
-      boundary_type == BoundaryConditions::BoundaryType::pressure)
+      boundary_type == BoundaryConditions::BoundaryType::pressure ||
+      boundary_type == BoundaryConditions::BoundaryType::neumann_traction)
     return true;
   else
     return false;
@@ -413,6 +414,16 @@ public:
    */
   void
   compute_inverse_diagonal(VectorType &diagonal) const;
+
+  /**
+   * @brief Store the values of the prescribed Neumann traction boundary condition to use it
+   * \f$
+   * (\boldsymbol{\sigma} \cdot \mathbf{n},\, \mathbf{v}) =
+   * (\mathrm{traction\_fn},\, \mathbf{v})\f$.
+   * in the residual assembly and Jacobian assembly of the matrix-free operator.
+   */
+  void
+  evaluate_prescribed_neumann_traction();
 
   /**
    * @brief Store relevant values of the vector of the last newton step to use it
@@ -792,6 +803,22 @@ protected:
    */
   Table<2, VectorizedArray<number>> face_target_pressure;
 
+  /**
+   * @brief Table with correct alignment for vectorization to store the values
+   * of the prescribed Neumann traction boundary condition values.
+   *
+   * \f$
+   * (\boldsymbol{\sigma} \cdot \mathbf{n},\, \mathbf{v}) =
+   * (\mathrm{traction\_fn},\, \mathbf{v})\f$.
+   * assembled \f$(\mathbf{v},\mathrm{traction\_fn})\f$ on boundary faces where
+   * Neumann traction is prescribed. our test function v involves both velocity
+   * and pressure components. Hence, the prescribed_neumann_traction has
+   * \f$\dim\f$ components corresponding to velocity and 1 component
+   * corresponding to pressure (which is dummy), totalling \f$\dim+1\f$
+   * components.
+   */
+  Table<2, Tensor<1, dim + 1, VectorizedArray<number>>>
+    prescribed_neumann_traction;
 
   /**
    * @brief Table with correct alignment for vectorization to store the values