+++ source/solvers/fluid_dynamics_matrix_free.cc
@@ -1065,6 +1065,11 @@ MFNavierStokesPreconditionGMGBase<dim>::reinit(
               /*The function weak boundary condition is implemented in the
                * operators*/
             }
+          else if (type == BoundaryConditions::BoundaryType::neumann_traction)
+            {
+              /*The Neumann traction boundary condition is implemented in the
+               * operators*/
+            }
           else if (type == BoundaryConditions::BoundaryType::partial_slip)
             {
               AssertThrow(
@@ -1620,6 +1625,12 @@ MFNavierStokesPreconditionGMGBase<dim>::reinit(
                 {
                   /*The directional do-nothing boundary condition is implemented
                    * in the operators*/
+                }
+              else if (type ==
+                       BoundaryConditions::BoundaryType::neumann_traction)
+                {
+                  /*The Neumann traction boundary condition is implemented in
+                   * the operators*/
                 }
               else if (type == BoundaryConditions::BoundaryType::noslip ||
                        type == BoundaryConditions::BoundaryType::function)
@@ -3312,6 +3323,13 @@ FluidDynamicsMatrixFree<dim>::assemble_system_rhs()
 
   this->system_operator->evaluate_non_linear_term_and_calculate_tau(
     this->evaluation_point);
+
+  /*
+    (σ . n , v)  = (traction_fn, v) on Γ_N. computes traction vector for the
+    prescribed Neumann traction σ = 2με(u)− pI
+  */
+
+  this->system_operator->evaluate_prescribed_neumann_traction();
 
   this->system_operator->evaluate_residual(this->system_rhs,
                                            this->evaluation_point);