+++ source/solvers/navier_stokes_base.cc
@@ -2107,6 +2107,11 @@ NavierStokesBase<dim, VectorType, DofsType>::define_zero_constraints()
           /*The directional do-nothing boundary condition is implemented
            * in the matrix-based assemblers and the matrix-free operators*/
         }
+      else if (type == BoundaryConditions::BoundaryType::neumann_traction)
+        {
+          /*The Neumann traction boundary condition is implemented
+           * in the  matrix-based assemblers and the matrix-free operators*/
+        }
       else if (type == BoundaryConditions::BoundaryType::none)
         {
           /*Default boundary condition*/
@@ -2256,7 +2261,8 @@ NavierStokesBase<dim, VectorType, DofsType>::update_boundary_conditions()
       if (type == BoundaryConditions::BoundaryType::function ||
           type == BoundaryConditions::BoundaryType::function_weak ||
           type == BoundaryConditions::BoundaryType::pressure ||
-          type == BoundaryConditions::BoundaryType::partial_slip)
+          type == BoundaryConditions::BoundaryType::partial_slip ||
+          type == BoundaryConditions::BoundaryType::neumann_traction)
         {
           this->simulation_parameters.boundary_conditions
             .navier_stokes_functions.at(id)
@@ -2270,6 +2276,9 @@ NavierStokesBase<dim, VectorType, DofsType>::update_boundary_conditions()
           this->simulation_parameters.boundary_conditions
             .navier_stokes_functions.at(id)
             ->p.set_time(time);
+          this->simulation_parameters.boundary_conditions
+            .navier_stokes_functions.at(id)
+            ->traction_fn.set_time(time);
         }
     }
   this->define_non_zero_constraints();