+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to the Lethe project will be documented in this file.
 The changelog for the previous releases of Lethe are located in the release_notes folder.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Master] - 2026/06/30
+
+### Added
+
+- MINOR Description (#1965).
+This PR implements Neumann traction boundary condition assembler for both matrix-based and matrix-free solvers to solve for the incompressible Navier-Stokes equations. The feature has been tested with the `method of manufactured solutions - 2d problem` adapted to have Neumann boundary condition on the top edge of the domain. Finally, a test is added for both Matrix-based and Matrix-free Navier-Stokes solvers solving the adapted mms_2d_problem with the `mms_2d_fe2_neumann_traction_navierstokes.prm` to keep track of its changes in the future. [#1965](https://github.com/chaos-polymtl/lethe/pull/1965)
+
 ## [Master] - 2026/06/28
 
 ### Changed
@@ -64,7 +71,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - MINOR The GCMG (global-coarsening multigrid) preconditioner of the VANS matrix-free solver was building the particle-fluid force transfer twice and never building the particle-fluid drag transfer (copy-paste bug). This left the drag contribution missing on coarse multigrid levels. The correct transfer is now built for each auxiliary field. [#1991](https://github.com/chaos-polymtl/lethe/pull/1991)
 
 ## [Sample] - YYYY/MM/DD
-
 ### Added
 
 - MAJOR/MINOR/PATCH Description (#PR).